+++ .idea/libraries/Gradle__android_arch_lifecycle_livedata_core_1_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.lifecycle:livedata-core-1.1.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/9a5d4d8616b05a4d563843c23df33689/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/9a5d4d8616b05a4d563843c23df33689/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/055e13402bc35f59ec49ad1f0f541eb4/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/livedata-core-1.1.0.aar/055e13402bc35f59ec49ad1f0f541eb4/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
