+++ .idea/libraries/Gradle__android_arch_lifecycle_runtime_1_1_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: android.arch.lifecycle:runtime-1.1.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/7a4983cc25d08b08ea03734cc79bc706/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/7a4983cc25d08b08ea03734cc79bc706/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/390cac4dbcf65d10798e1688487e5a72/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runtime-1.1.0.aar/390cac4dbcf65d10798e1688487e5a72/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
