+++ .idea/libraries/Gradle__com_android_support_animated_vector_drawable_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:animated-vector-drawable-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/8b6b87096b8d5588a8d45d840d3b1537/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/8b6b87096b8d5588a8d45d840d3b1537/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/92a9dde326188e613fdf2ddad0c96734/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/animated-vector-drawable-27.1.1.aar/92a9dde326188e613fdf2ddad0c96734/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
