+++ .idea/libraries/Gradle__com_android_support_appcompat_v7_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:appcompat-v7-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/1d95cd3ac07c8418d6dd99f73d68da80/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/1d95cd3ac07c8418d6dd99f73d68da80/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/0244a026c6b8a34af3b87af04f0b298e/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/appcompat-v7-27.1.1.aar/0244a026c6b8a34af3b87af04f0b298e/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
