+++ .idea/libraries/Gradle__com_android_support_design_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:design-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/5572b0f09860c483ad3370d76cdaaeb0/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/5572b0f09860c483ad3370d76cdaaeb0/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/b5592c7df2af272364fc6bc331ab12e8/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/design-27.1.1.aar/b5592c7df2af272364fc6bc331ab12e8/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
