+++ .idea/libraries/Gradle__com_android_support_support_core_ui_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-core-ui-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/e2f7b5713b3bb1675f28e8b35ad13945/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/e2f7b5713b3bb1675f28e8b35ad13945/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/bf5690c2f8a5ac48e6e44d481b293ffe/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-core-ui-27.1.1.aar/bf5690c2f8a5ac48e6e44d481b293ffe/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
