+++ .idea/libraries/Gradle__com_android_support_support_fragment_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-fragment-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/060c10d186947450514ebb453c3ebcb8/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/060c10d186947450514ebb453c3ebcb8/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/c6f44de982f1457f5ec28a2dc14a57a3/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-fragment-27.1.1.aar/c6f44de982f1457f5ec28a2dc14a57a3/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
