+++ .idea/libraries/Gradle__com_android_support_support_media_compat_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-media-compat-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/38dacd65ea717bfa8b6b9cd52839c217/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/38dacd65ea717bfa8b6b9cd52839c217/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/0a5d5cca251bfef8e83e19209550d8a2/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-media-compat-27.1.1.aar/0a5d5cca251bfef8e83e19209550d8a2/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
