+++ .idea/libraries/Gradle__com_android_support_support_vector_drawable_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:support-vector-drawable-27.1.1">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/f0d2256bc4f28be2c1879ca2cfde7791/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/f0d2256bc4f28be2c1879ca2cfde7791/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/aaed6cff3d992007a932c11a4504b412/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/support-vector-drawable-27.1.1.aar/aaed6cff3d992007a932c11a4504b412/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
