+++ .idea/libraries/Gradle__com_android_support_test_espresso_espresso_core_3_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test.espresso:espresso-core-3.0.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/bc1aa4992ba05686c214f92e7052dfcf/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/bc1aa4992ba05686c214f92e7052dfcf/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/77f72deaabb4cc2f7ceaa0a52d6885a8/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/77f72deaabb4cc2f7ceaa0a52d6885a8/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
