+++ .idea/libraries/Gradle__com_android_support_test_espresso_espresso_idling_resource_3_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test.espresso:espresso-idling-resource-3.0.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/12e39574278a370a1c2c0cbf101b2ffe/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/12e39574278a370a1c2c0cbf101b2ffe/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/9b04cd2f44c61e16b20cd46aa4fd40a1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/9b04cd2f44c61e16b20cd46aa4fd40a1/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
