+++ .idea/libraries/Gradle__com_android_support_test_monitor_1_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test:monitor-1.0.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/e0b5129299a32960d576b19f5ca8261f/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/e0b5129299a32960d576b19f5ca8261f/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/0763a2f2fbce3d7bade0f3e5cfa32bf4/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/0763a2f2fbce3d7bade0f3e5cfa32bf4/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
