+++ .idea/libraries/Gradle__com_android_support_test_runner_1_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test:runner-1.0.2">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/42b52806b174db2ab7d7931b56c2b283/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/42b52806b174db2ab7d7931b56c2b283/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/d0f5352462680edb5a3797598763436b/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/d0f5352462680edb5a3797598763436b/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
