+++ .idea/libraries/Gradle__com_android_support_transition_27_1_1.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support:transition-27.1.1">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/73183375b1bcd13b842f9cc334387ae7/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/73183375b1bcd13b842f9cc334387ae7/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/bffa197ceba6d1dc65c4425f2326ac82/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/transition-27.1.1.aar/bffa197ceba6d1dc65c4425f2326ac82/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
