+++ .idea/libraries/Gradle__com_google_android_gms_play_services_base_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.android.gms:play-services-base-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-11.8.0.aar/6299e866ae6d981088c13fe7f3932eeb/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-11.8.0.aar/6299e866ae6d981088c13fe7f3932eeb/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-11.8.0.aar/c8aed2956d643532d9c61691fcff5ef7/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-11.8.0.aar/c8aed2956d643532d9c61691fcff5ef7/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
