+++ .idea/libraries/Gradle__com_google_android_gms_play_services_base_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.android.gms:play-services-base-license-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-license-11.8.0.aar/8bb59c8c8fbfbbe8807660ae170ce8f0/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-license-11.8.0.aar/8bb59c8c8fbfbbe8807660ae170ce8f0/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-license-11.8.0.aar/d0017f617b5e81411e8cfb24adea9e33/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-base-license-11.8.0.aar/d0017f617b5e81411e8cfb24adea9e33/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
