+++ .idea/libraries/Gradle__com_google_android_gms_play_services_basement_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.android.gms:play-services-basement-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-11.8.0.aar/54bae339c801914fe48dd763a3b05f22/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-11.8.0.aar/54bae339c801914fe48dd763a3b05f22/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-11.8.0.aar/9fefc4b96ccc6cbbb9b74f77628fe024/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-11.8.0.aar/9fefc4b96ccc6cbbb9b74f77628fe024/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
