+++ .idea/libraries/Gradle__com_google_android_gms_play_services_basement_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.android.gms:play-services-basement-license-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-license-11.8.0.aar/b385ee495d20e6808ab0faefb9811d01/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-license-11.8.0.aar/b385ee495d20e6808ab0faefb9811d01/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-license-11.8.0.aar/5777b4ff75fbad9a8f6fc76748665793/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-basement-license-11.8.0.aar/5777b4ff75fbad9a8f6fc76748665793/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
