+++ .idea/libraries/Gradle__com_google_android_gms_play_services_tasks_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.android.gms:play-services-tasks-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-11.8.0.aar/bde88a90f6c85ea691982382fe25a266/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-11.8.0.aar/bde88a90f6c85ea691982382fe25a266/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-11.8.0.aar/9c3c6f8db4c1d8c3d95265f4e33dba72/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-11.8.0.aar/9c3c6f8db4c1d8c3d95265f4e33dba72/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
