+++ .idea/libraries/Gradle__com_google_android_gms_play_services_tasks_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.android.gms:play-services-tasks-license-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-license-11.8.0.aar/60de22232e26d00ab11fd3a9f86dd531/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-license-11.8.0.aar/60de22232e26d00ab11fd3a9f86dd531/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-license-11.8.0.aar/f545de0b25a80d0f2ba378d875bbb3b2/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/play-services-tasks-license-11.8.0.aar/f545de0b25a80d0f2ba378d875bbb3b2/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
