+++ .idea/libraries/Gradle__com_google_firebase_firebase_analytics_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-analytics-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-11.8.0.aar/587bc9f22514f087c70d4f5b83f5f25e/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-11.8.0.aar/587bc9f22514f087c70d4f5b83f5f25e/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-11.8.0.aar/4817c2af11a07e7f879b7fc414be750e/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-11.8.0.aar/4817c2af11a07e7f879b7fc414be750e/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
