+++ .idea/libraries/Gradle__com_google_firebase_firebase_analytics_impl_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-analytics-impl-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-impl-11.8.0.aar/8cdd3a0562a1398e62a79c675c494342/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-impl-11.8.0.aar/8cdd3a0562a1398e62a79c675c494342/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-impl-11.8.0.aar/a223669e5f2d2177926d4145c7412bde/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-impl-11.8.0.aar/a223669e5f2d2177926d4145c7412bde/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
