+++ .idea/libraries/Gradle__com_google_firebase_firebase_analytics_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-analytics-license-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-license-11.8.0.aar/b05ae84f650f2634b6f638b334d88042/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-license-11.8.0.aar/b05ae84f650f2634b6f638b334d88042/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-license-11.8.0.aar/08aa4f88c95e49c998ed37d780b1bd89/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-analytics-license-11.8.0.aar/08aa4f88c95e49c998ed37d780b1bd89/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
