+++ .idea/libraries/Gradle__com_google_firebase_firebase_auth_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-auth-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-11.8.0.aar/b5aa9b2c9390a8c72c46c71d9d6564e3/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-11.8.0.aar/b5aa9b2c9390a8c72c46c71d9d6564e3/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-11.8.0.aar/6a2d11e0f7e8812add9bc4cf0c84b028/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-11.8.0.aar/6a2d11e0f7e8812add9bc4cf0c84b028/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
