+++ .idea/libraries/Gradle__com_google_firebase_firebase_auth_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-auth-license-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-license-11.8.0.aar/49b0adf89b1537c08ab103c96f75b786/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-license-11.8.0.aar/49b0adf89b1537c08ab103c96f75b786/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-license-11.8.0.aar/b141f1ee71799838037c93bc0bc5b79f/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-auth-license-11.8.0.aar/b141f1ee71799838037c93bc0bc5b79f/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
