+++ .idea/libraries/Gradle__com_google_firebase_firebase_common_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-common-license-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-common-license-11.8.0.aar/c822a1b722b60d5ac8148313f21e8d41/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-common-license-11.8.0.aar/c822a1b722b60d5ac8148313f21e8d41/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-common-license-11.8.0.aar/718f352ec628c432f92cdbf09f2ced6f/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-common-license-11.8.0.aar/718f352ec628c432f92cdbf09f2ced6f/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
