+++ .idea/libraries/Gradle__com_google_firebase_firebase_core_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-core-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-core-11.8.0.aar/57f0db1bc4fc39e0f0470731825374b5/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-core-11.8.0.aar/57f0db1bc4fc39e0f0470731825374b5/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-core-11.8.0.aar/e55ad3dbc71c491d2da16166dab790dc/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-core-11.8.0.aar/e55ad3dbc71c491d2da16166dab790dc/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
