+++ .idea/libraries/Gradle__com_google_firebase_firebase_iid_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-iid-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-11.8.0.aar/43c3653709678a5aa4dc31de9e99f113/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-11.8.0.aar/43c3653709678a5aa4dc31de9e99f113/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-11.8.0.aar/b89901f833ad3195458898cf7562be98/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-11.8.0.aar/b89901f833ad3195458898cf7562be98/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
