+++ .idea/libraries/Gradle__com_google_firebase_firebase_iid_license_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-iid-license-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-license-11.8.0.aar/f365d7f052213283f65338423ebf438e/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-license-11.8.0.aar/f365d7f052213283f65338423ebf438e/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-license-11.8.0.aar/8d236086d9b4444e3ebf4492c5ad0a6c/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-iid-license-11.8.0.aar/8d236086d9b4444e3ebf4492c5ad0a6c/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
