+++ .idea/libraries/Gradle__com_google_firebase_firebase_storage_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-storage-11.8.0">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-11.8.0.aar/a645d200a24883b1497c8634c87aad59/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-11.8.0.aar/a645d200a24883b1497c8634c87aad59/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-11.8.0.aar/82a95b327c95c1243a6952bdd58cb63b/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-11.8.0.aar/82a95b327c95c1243a6952bdd58cb63b/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
