+++ .idea/libraries/Gradle__com_google_firebase_firebase_storage_common_11_8_0.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.google.firebase:firebase-storage-common-11.8.0">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-common-11.8.0.aar/63a1687305e76cda772cdb1f8410e229/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-common-11.8.0.aar/63a1687305e76cda772cdb1f8410e229/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-common-11.8.0.aar/2e3479fa7b6c598478089e7017b0d8f7/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/firebase-storage-common-11.8.0.aar/2e3479fa7b6c598478089e7017b0d8f7/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
